requirements.txt
flask==1.1.1
flask-restful==0.3.7
flask-cors==3.0.8
Flask-Mail==0.9.1
pymongo==3.8.0
dnspython
gunicorn
git+https://github.com/FaizanMostafa/robin_stocks.git
uuid==1.30
bcrypt==3.1.7
requests==2.24.0
python-dotenv==0.15.0
yahoofinancials==1.6
cryptography==3.4.6
Flask-JWT-Extended==4.1.0
APScheduler==3.6.1
pyportfolioopt==1.4.1
pandas-datareader==0.9.0
pandas==0.25.1
numpy==1.19.3
onesignal-sdk==2.0.0
yahoo-fin==0.8.8
pytz